requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.2
Django==4.2.4
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
Flask==2.2.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
packaging==23.1
Pillow==9.4.0
pluggy==1.2.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
rest-framework-simplejwt==0.0.1
rfc3986==1.5.0
rich==12.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
sqlparse==0.4.3
tabulate==0.9.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==2.0.4
Werkzeug==2.2.2
